+++ pysys-examples/internal/testcases/PySys_internal_002/run.py
@@ -12,6 +12,11 @@ class PySysTest(BaseTest):
 				self.log.info("%s = %s", attr, eval("self.proj.%s" % attr))
 		stdoutHandler.setFormatter(PROJECT.formatters.stdout)
 		
+		# good opportunity to check that default project props are set
+		self.assertThat('%s != ""', repr(self.proj.os)) # renamed from osfamily by config
+		self.assertThat('%s != ""', repr(self.proj.root))
+		self.assertThat('%s != ""', repr(self.proj.hostname))
+		
 	def validate(self):
 		self.assertTrue(self.proj.lib == 'lib_%s_1.0.so'%OSFAMILY)
 		self.assertTrue(self.proj.library == 'jstore1.0.jar')